+++ tools/yulPhaser/SimulationRNG.h
@@ -14,6 +14,7 @@
 	You should have received a copy of the GNU General Public License
 	along with solidity.  If not, see <http://www.gnu.org/licenses/>.
 */
+// SPDX-License-Identifier: GPL-3.0
 
 #pragma once
 
@@ -38,8 +39,8 @@ class SimulationRNG
 {
 public:
 	static bool bernoulliTrial(double _successProbability);
-	static uint32_t uniformInt(uint32_t _min, uint32_t _max);
-	static uint32_t binomialInt(uint32_t _numTrials, double _successProbability);
+	static size_t uniformInt(size_t _min, size_t _max);
+	static size_t binomialInt(size_t _numTrials, double _successProbability);
 
 	/// Resets generator to a known state given by the @a seed. Given the same seed, a fixed
 	/// sequence of calls to the members generating random values is guaranteed to produce the
